Restructured training and began implementing Megatron-LM training
