+++ examples/llama_1b.sh
@@ -1,4 +1,4 @@
-accelerate launch --config_file configs/bf16_z0_2gpu.yaml pretrain.py \
+accelerate launch --config_file configs/deepspeed.yaml pretrain.py \
     --epochs 5 \
     --data_path "." \
     --config_path "TinyLlama/TinyLlama-1.1B-Chat-v1.0" \
